Rename userinput to dialog
