+++ src/cmd/config.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	"github.com/git-town/git-town/v7/src/cli"
+	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/git-town/git-town/v7/src/git"
-	"github.com/git-town/git-town/v7/src/userinput"
 	"github.com/spf13/cobra"
 )
 
@@ -176,7 +176,7 @@ var updatePrennialBranchesCommand = &cobra.Command{
 	Short: "Prompts to update your perennial branches",
 	Long:  `Prompts to update your perennial branches`,
 	Run: func(cmd *cobra.Command, args []string) {
-		err := userinput.ConfigurePerennialBranches(prodRepo)
+		err := dialog.ConfigurePerennialBranches(prodRepo)
 		if err != nil {
 			cli.Exit(err)
 		}
@@ -235,11 +235,11 @@ var setupConfigCommand = &cobra.Command{
 	Use:   "setup",
 	Short: "Prompts to setup your Git Town configuration",
 	Run: func(cmd *cobra.Command, args []string) {
-		err := userinput.ConfigureMainBranch(prodRepo)
+		err := dialog.ConfigureMainBranch(prodRepo)
 		if err != nil {
 			cli.Exit(err)
 		}
-		err = userinput.ConfigurePerennialBranches(prodRepo)
+		err = dialog.ConfigurePerennialBranches(prodRepo)
 		if err != nil {
 			cli.Exit(err)
 		}
