+++ src/cmd/kill.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 
 	"github.com/git-town/git-town/v7/src/cli"
+	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/git-town/git-town/v7/src/git"
 	"github.com/git-town/git-town/v7/src/runstate"
 	"github.com/git-town/git-town/v7/src/steps"
-	"github.com/git-town/git-town/v7/src/userinput"
 	"github.com/spf13/cobra"
 )
 
@@ -74,7 +74,7 @@ func createKillConfig(args []string, repo *git.ProdRepo) (killConfig, error) {
 		return result, err
 	}
 	if result.isTargetBranchLocal {
-		err = userinput.EnsureKnowsParentBranches([]string{result.targetBranch}, repo)
+		err = dialog.EnsureKnowsParentBranches([]string{result.targetBranch}, repo)
 		if err != nil {
 			return result, err
 		}
