+++ src/cmd/new_pull_request.go
@@ -2,11 +2,11 @@ package cmd
 
 import (
 	"github.com/git-town/git-town/v7/src/cli"
+	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/git-town/git-town/v7/src/git"
 	"github.com/git-town/git-town/v7/src/hosting"
 	"github.com/git-town/git-town/v7/src/runstate"
 	"github.com/git-town/git-town/v7/src/steps"
-	"github.com/git-town/git-town/v7/src/userinput"
 	"github.com/spf13/cobra"
 )
 
@@ -83,7 +83,7 @@ func createNewPullRequestConfig(repo *git.ProdRepo) (newPullRequestConfig, error
 	if err != nil {
 		return newPullRequestConfig{}, err
 	}
-	err = userinput.EnsureKnowsParentBranches([]string{initialBranch}, repo)
+	err = dialog.EnsureKnowsParentBranches([]string{initialBranch}, repo)
 	if err != nil {
 		return newPullRequestConfig{}, err
 	}
