+++ src/cmd/shared.go
@@ -4,11 +4,11 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/git-town/git-town/v7/src/git"
 	"github.com/git-town/git-town/v7/src/hosting"
 	"github.com/git-town/git-town/v7/src/runstate"
 	"github.com/git-town/git-town/v7/src/steps"
-	"github.com/git-town/git-town/v7/src/userinput"
 )
 
 // These variables represent command-line flags.
@@ -29,7 +29,7 @@ var (
 const dryRunFlagDescription = "Print the commands but don't run them"
 
 func validateIsConfigured(repo *git.ProdRepo) error {
-	err := userinput.EnsureIsConfigured(repo)
+	err := dialog.EnsureIsConfigured(repo)
 	if err != nil {
 		return err
 	}
@@ -75,7 +75,7 @@ func handleUnfinishedState(repo *git.ProdRepo, driver hosting.Driver) (quit bool
 	if runState == nil || !runState.IsUnfinished() {
 		return false, nil
 	}
-	response, err := userinput.AskHowToHandleUnfinishedRunState(
+	response, err := dialog.AskHowToHandleUnfinishedRunState(
 		runState.Command,
 		runState.UnfinishedDetails.EndBranch,
 		runState.UnfinishedDetails.EndTime,
@@ -85,10 +85,10 @@ func handleUnfinishedState(repo *git.ProdRepo, driver hosting.Driver) (quit bool
 		return quit, err
 	}
 	switch response {
-	case userinput.ResponseTypeDiscard:
+	case dialog.ResponseTypeDiscard:
 		err = runstate.Delete(repo)
 		return false, err
-	case userinput.ResponseTypeContinue:
+	case dialog.ResponseTypeContinue:
 		hasConflicts, err := repo.Silent.HasConflicts()
 		if err != nil {
 			return false, err
@@ -97,10 +97,10 @@ func handleUnfinishedState(repo *git.ProdRepo, driver hosting.Driver) (quit bool
 			return false, fmt.Errorf("you must resolve the conflicts before continuing")
 		}
 		return true, runstate.Execute(runState, repo, driver)
-	case userinput.ResponseTypeAbort:
+	case dialog.ResponseTypeAbort:
 		abortRunState := runState.CreateAbortRunState()
 		return true, runstate.Execute(&abortRunState, repo, driver)
-	case userinput.ResponseTypeSkip:
+	case dialog.ResponseTypeSkip:
 		skipRunState := runState.CreateSkipRunState()
 		return true, runstate.Execute(&skipRunState, repo, driver)
 	default:
