+++ src/dialog/branch_helpers.go
@@ -1,4 +1,4 @@
-package userinput
+package dialog
 
 import (
 	"fmt"
