+++ src/dialog/doc.go
@@ -1,0 +1,2 @@
+// Package dialog allows the user to enter configuration data via CLI dialogs and prompts.
+package dialog
