+++ src/dialog/parent_branches.go
@@ -1,4 +1,4 @@
-package userinput
+package dialog
 
 import (
 	"fmt"
