+++ src/dialog/squash_commit_author.go
@@ -1,4 +1,4 @@
-package userinput
+package dialog
 
 import (
 	"fmt"
