--- src/userinput/doc.go
@@ -1,2 +1,0 @@
-// Package userinput allows the user to enter configuration data via CLI dialogs and prompts.
-package userinput
